+++ fixtures/gists-app/app/routes/gists.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
 import { Link, useRouteData, useLocationPending } from "@remix-run/react";
+
 import Shared from "../components/Shared";
 
 export function headers() {