+++ fixtures/gists-app/app/routes/index.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { Link } from "@remix-run/react";
+
 import Shared from "../components/Shared";
 
 export function meta() {